+++ common/db/src/main/java/is/codion/common/db/connection/DatabaseConnection.java
@@ -82,15 +82,17 @@ public interface DatabaseConnection extends AutoCloseable {
 
 	/**
 	 * Performs a commit and ends the current transaction
+	 * @throws SQLException in case the commit failed
 	 * @throws IllegalStateException in case a transaction is not open
 	 */
-	void commitTransaction();
+	void commitTransaction() throws SQLException;
 
 	/**
 	 * Performs a rollback and ends the current transaction
+	 * @throws SQLException in case the rollback failed
 	 * @throws IllegalStateException in case a transaction is not open
 	 */
-	void rollbackTransaction();
+	void rollbackTransaction() throws SQLException;
 
 	/**
 	 * Performs a commit