+++ demos/manual/src/main/java/is/codion/framework/demos/manual/common/db/Demo.java
@@ -77,7 +77,7 @@ public final class Demo {
 		// end::connection[]
 	}
 
-	private void databaseConnection() throws DatabaseException {
+	private void databaseConnection() throws DatabaseException, SQLException {
 		// tag::databaseConnection[]
 		Database.DATABASE_URL.set("jdbc:h2:mem:h2db");